[chore] #18  data <-domain 의존성 분리를 위해 dto와 로직변경
