+++ app/src/main/java/com/sopt/now/data/dto/request/RequestSignUpDto.kt
@@ -1,7 +1,5 @@
 package com.sopt.now.data.dto.request
 
-import com.sopt.now.data.model.User
-import com.sopt.now.domain.entity.request.AuthRequestModel
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -15,16 +13,4 @@ data class RequestSignUpDto(
     val nickname: String,
     @SerialName("phone")
     val phone: String,
-) {
-    fun toUserWithUserId(userid: String): User =
-        User(
-            id = authenticationId,
-            password = password,
-            nickname = nickname,
-            phoneNumber = phone,
-            userId = userid
-        )
-
-    fun toAuthRequestModel(): AuthRequestModel =
-        AuthRequestModel(authenticationId, password, nickname, phone)
-}
+)
